Successfully moved Auth routes to alt file
