+++ src/index.js
@@ -4,36 +4,10 @@
 
 const express = require('express'); // To build an application server or API
 const app = express();
-const pgp = require('pg-promise')(); // To connect to the Postgres DB from the node server
 const bodyParser = require('body-parser');
 const session = require('express-session'); // To set the session object. To store or access session data, use the `req.session`, which is (generally) serialized as JSON by the store.
 const bcrypt = require('bcrypt'); //  To hash passwords
 const axios = require('axios'); // To make HTTP requests from our server. We'll learn more about it in Part B.
-
-// *****************************************************
-// <!-- Section 2 : Connect to DB -->
-// *****************************************************
-
-// database configuration
-const dbConfig = {
-  host: 'db', // the database server
-  port: 5432, // the database port
-  database: process.env.POSTGRES_DB, // the database name
-  user: process.env.POSTGRES_USER, // the user account to connect with
-  password: process.env.POSTGRES_PASSWORD, // the password of the user account
-};
-
-const db = pgp(dbConfig);
-
-// test your database
-db.connect()
-  .then(obj => {
-    console.log('Database connection successful'); // you can view this message in the docker compose logs
-    obj.done(); // success, release the connection;
-  })
-  .catch(error => {
-    console.log('ERROR:', error.message || error);
-  });
 
 // *****************************************************
 // <!-- Section 3 : App Settings -->
@@ -58,13 +32,13 @@ app.use(
 );
 
 // Importing the post routes
-import authRoutes from "./js/auth.routes";
+const authRoutes = require("./js/auth.routes");
 
 // *****************************************************
 // <!-- Section 4 : API Routes -->
 // *****************************************************
 
-// TODO - Include your API routes here
+//! Page rendering routes
 
 app.get('/', (req, res) => {
     res.redirect('/login')
@@ -85,7 +59,11 @@ app.get('/logout', (req, res) => {
     });
 });
 
-app.use('/', authRoutes);
+//! API routes
+
+//? Route for all authroutes is localhost:3000/auth + the route located on auth.routes.js
+// For example, localhost:3000/auth/register will route to the register page
+app.use('/auth', authRoutes);
 
 app.post('/login', async (req, res) => {
