+++ src/views/pages/login.ejs
@@ -6,7 +6,7 @@
   <div class="container">
     <!-- THE REST OF YOUR HTML CODE GOES HERE -->
     <h1><centre>Login</centre></h1>
-    <form action="/login" method="POST">
+    <form action="/auth/login" method="POST">
       <div class="row mb-3">
         <label for="colFormLabel" class="col-sm-2 col-form-label">Username</label>
         <div class="col-sm-10">
